support for disable classes added.
With this change each element that contains the `data-disable-with`
attribute can provides an attribute called `data-disable-class`
which would be a string of classes to be added to the element
after the element disabled. For example using this feature it
would be possible to pass `disabled` css class to a bootstrap button.