Update README.md
Adds usage for modern versions of Rails. also attempts to alert the developer to common pitfalls, which are more common now that there are more steps to get jQuery installed through Webpack

this should clear up a lot of confusion and save a lot of debugging on new apps when trying to configure/install jquery for the first time.